pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "founder-product-feedback-roadmap-os"
version = "0.1.0"
description = "Deterministic founder-facing product feedback and roadmap prioritization operating system."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Shubham Singh" }
]
dependencies = [
  "pandas>=2.0",
  "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
